+++ MWSE/StackLua.cpp
@@ -18,15 +18,15 @@ namespace mwse {
 			// Functions for pushing values.
 			//
 
-			state["mwse"]["stack"]["pushShort"] = [](double value) {
+			state["mwse"]["stack"]["pushShort"] = [](short value) {
 				Stack::getInstance().pushShort(value);
 			};
 
-			state["mwse"]["stack"]["pushLong"] = [](double value) {
+			state["mwse"]["stack"]["pushLong"] = [](int value) {
 				Stack::getInstance().pushLong(value);
 			};
 
-			state["mwse"]["stack"]["pushFloat"] = [](double value) {
+			state["mwse"]["stack"]["pushFloat"] = [](float value) {
 				Stack::getInstance().pushFloat(value);
 			};
 
@@ -35,7 +35,7 @@ namespace mwse {
 			};
 
 			state["mwse"]["stack"]["pushObject"] = [](sol::object value) {
-				Stack::getInstance().pushLong((long)value.as<TES3::BaseObject*>());
+				Stack::getInstance().pushLong((DWORD)value.as<TES3::BaseObject*>());
 			};
 
 			//
